+++ spec/spec_helper.rb
@@ -1,0 +1,4 @@
+require 'chefspec'
+require 'chefspec/server'
+
+at_exit { ChefSpec::Coverage.report! }